Split main_test.py into a few modules
